hack and restart design of hivemind stuff
